requirements.txt
fastapi>=0.95.2,<1.0.0
uvicorn>=0.15.0,<1.0.0
motor>=3.0.0,<4.0.0
pydantic[dotenv]>=1.10.0,<2.0.0
fastapi-jwt-auth>=0.5.0,<1.0.0
python-dotenv>=0.21.0,<1.0.0